+++ src/actions/synchronization/taskunifier/TaskUnifierNoteActions.js
@@ -24,9 +24,7 @@ export function synchronizeNotes() {
             if (notesToAdd.length > 0) {
                 const result = await dispatch(addRemoteNotes(notesToAdd));
 
-                for (let note of result) {
-                    await dispatch(updateNote(note, { loaded: true, skipUpdateMiddleware: true }));
-                }
+                await dispatch(updateNote(result, { loaded: true, skipUpdateMiddleware: true }));
             }
         }
 
@@ -39,40 +37,47 @@ export function synchronizeNotes() {
                 await dispatch(deleteRemoteNotes(notesToDelete));
             }
 
-            for (let note of notesToDelete) {
-                await dispatch(deleteNote(note.id));
-            }
+            await dispatch(deleteNote(notesToDelete.map(note => note.id)));
         }
 
         notes = getNotes(getState());
 
         const lastSync = settings.lastSynchronizationDate ? moment(settings.lastSynchronizationDate) : null;
+
+        const notesToAdd = [];
+        const notesToUpdate = [];
         const remoteNotes = await dispatch(getRemoteNotes(lastSync));
 
         for (let remoteNote of remoteNotes) {
             const localNote = notes.find(note => note.refIds.taskunifier === remoteNote.refIds.taskunifier);
 
             if (!localNote) {
-                await dispatch(addNote(remoteNote, { keepRefIds: true }));
+                notesToAdd.push(remoteNote);
             } else {
                 if (moment(remoteNote.updateDate).diff(moment(localNote.updateDate)) > 0) {
-                    await dispatch(updateNote(merge(localNote, remoteNote), { loaded: true, skipUpdateMiddleware: true }));
+                    notesToUpdate.push(merge(localNote, remoteNote));
                 }
             }
         }
 
+        await dispatch(addNote(notesToAdd, { keepRefIds: true }));
+        await dispatch(updateNote(notesToUpdate, { loaded: true, skipUpdateMiddleware: true }));
+
         notes = getNotes(getState());
 
         {
+            const notesToDelete = [];
             const remoteDeletedNotes = await dispatch(getRemoteDeletedNotes(lastSync));
 
             for (let remoteDeletedNote of remoteDeletedNotes) {
                 const localNote = notes.find(note => note.refIds.taskunifier === remoteDeletedNote.id);
 
                 if (localNote) {
-                    await dispatch(deleteNote(localNote.id, { force: true }));
+                    notesToDelete.push(localNote);
                 }
             }
+
+            await dispatch(deleteNote(notesToDelete.map(note => note.id), { force: true }));
         }
 
         notes = getNotes(getState());
@@ -84,9 +89,7 @@ export function synchronizeNotes() {
                 await dispatch(editRemoteNotes(notesToUpdate, remoteNotes));
             }
 
-            for (let note of notesToUpdate) {
-                await dispatch(updateNote(note, { loaded: true, skipUpdateMiddleware: true }));
-            }
+            await dispatch(updateNote(notesToUpdate, { loaded: true, skipUpdateMiddleware: true }));
         }
     };
 }