+++ src/actions/synchronization/toodledo/ToodledoFolderActions.js
@@ -21,9 +21,7 @@ export function synchronizeFolders() {
             const foldersToAddPromises = foldersToAdd.map(folder => dispatch(addRemoteFolder(folder)));
             const result = await Promise.all(foldersToAddPromises);
 
-            for (let folder of result) {
-                await dispatch(updateFolder(folder, { loaded: true, skipUpdateMiddleware: true }));
-            }
+            await dispatch(updateFolder(result, { loaded: true, skipUpdateMiddleware: true }));
         }
 
         folders = getFolders(getState());
@@ -33,9 +31,7 @@ export function synchronizeFolders() {
             const foldersToDeletePromises = foldersToDelete.map(folder => dispatch(deleteRemoteFolder(folder)));
             await Promise.all(foldersToDeletePromises);
 
-            for (let folder of foldersToDelete) {
-                await dispatch(deleteFolder(folder.id));
-            }
+            await dispatch(deleteFolder(foldersToDelete.map(folder => folder.id)));
         }
 
         folders = getFolders(getState());
@@ -45,26 +41,34 @@ export function synchronizeFolders() {
             const lastEditFolder = moment.unix(getToodledoAccountInfo(getState()).lastedit_folder);
 
             if (!lastSync || lastEditFolder.diff(lastSync) > 0) {
+                const foldersToAdd = [];
+                const foldersToUpdate = [];
+                const foldersToDelete = [];
                 const remoteFolders = await dispatch(getRemoteFolders());
 
                 for (let remoteFolder of remoteFolders) {
                     const localFolder = folders.find(folder => folder.refIds.toodledo === remoteFolder.refIds.toodledo);
 
                     if (!localFolder) {
-                        await dispatch(addFolder(remoteFolder, { keepRefIds: true }));
+                        foldersToAdd.push(remoteFolder);
                     } else {
-                        await dispatch(updateFolder(merge(localFolder, remoteFolder), { loaded: true, skipUpdateMiddleware: true }));
+                        foldersToUpdate.push(merge(localFolder, remoteFolder));
                     }
                 }
+
+                await dispatch(addFolder(foldersToAdd, { keepRefIds: true }));
+                await dispatch(updateFolder(foldersToUpdate, { loaded: true, skipUpdateMiddleware: true }));
 
                 folders = getFolders(getState());
 
                 // eslint-disable-next-line require-atomic-updates
                 for (let localFolder of filterByVisibleState(folders)) {
                     if (!remoteFolders.find(folder => folder.refIds.toodledo === localFolder.refIds.toodledo)) {
-                        await dispatch(deleteFolder(localFolder.id, { force: true }));
+                        foldersToDelete.push(localFolder);
                     }
                 }
+
+                await dispatch(deleteFolder(foldersToDelete.map(folder => folder.id), { force: true }));
             }
         }
 
@@ -75,9 +79,7 @@ export function synchronizeFolders() {
             const foldersToUpdatePromises = foldersToUpdate.map(folder => dispatch(editRemoteFolder(folder)));
             await Promise.all(foldersToUpdatePromises);
 
-            for (let folder of foldersToUpdate) {
-                await dispatch(updateFolder(folder, { loaded: true, skipUpdateMiddleware: true }));
-            }
+            await dispatch(updateFolder(foldersToUpdate, { loaded: true, skipUpdateMiddleware: true }));
         }
     };
 }